Create get_information.py
Scripts that reads a student card (Read.py) and searches the LDAP server for the corresponding st, which is the RFID tag number,and if it does find it, it'll dump info (Name, email, student grade) to a file that will be later on read by MyP13 App (C++). 
In short, this script is used to authenticate the user with his student's card.